+++ src/main/java/org/realityforge/bazel/depgen/config/JavaConfig.java
@@ -1,0 +1,20 @@
+package org.realityforge.bazel.depgen.config;
+
+import javax.annotation.Nullable;
+
+public final class JavaConfig
+{
+  @Nullable
+  private Boolean exportDeps;
+
+  @Nullable
+  public Boolean getExportDeps()
+  {
+    return exportDeps;
+  }
+
+  public void setExportDeps( @Nullable final Boolean exportDeps )
+  {
+    this.exportDeps = exportDeps;
+  }
+}